Feature/getting rid of warnings

+++ Example/Tests/freestyle/Styling/PXFontInfoTests.m
@@ -27,7 +27,7 @@
     NSArray *result = [PXFontEntry filterEntries:entries byStretch:[PXFontEntry indexFromStretchName:@"normal"]];
 
     XCTAssertNotNil(result, @"Expected an array back when filtering by stretch");
-    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %d", result.count);
+    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %lu", (unsigned long)result.count);
     PXFontEntry *item = result[0];
     XCTAssertEqualObjects(normal, item, @"Expected to find normal stretch font");
 }
@@ -42,7 +42,7 @@
     NSArray *result = [PXFontEntry filterEntries:entries byStretch:[PXFontEntry indexFromStretchName:@"normal"]];
 
     XCTAssertNotNil(result, @"Expected an array back when filtering by stretch");
-    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %d", result.count);
+    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %lu", (unsigned long)result.count);
     PXFontEntry *item = result[0];
     XCTAssertEqualObjects(condensed, item, @"Expected to find normal stretch font");
 }
@@ -58,7 +58,7 @@
     NSArray *result = [PXFontEntry filterEntries:entries byWeight:[PXFontEntry indexFromWeightName:@"normal"]];
 
     XCTAssertNotNil(result, @"Expected an array back when filtering by weight");
-    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %d", result.count);
+    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %lu", (unsigned long)result.count);
     PXFontEntry *item = result[0];
     XCTAssertEqualObjects(normal, item, @"Expected %@ but found %@", normal, item);
 }
@@ -73,7 +73,7 @@
     NSArray *result = [PXFontEntry filterEntries:entries byWeight:[PXFontEntry indexFromWeightName:@"normal"]];
 
     XCTAssertNotNil(result, @"Expected an array back when filtering by weight");
-    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %d", result.count);
+    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %lu", (unsigned long)result.count);
     PXFontEntry *item = result[0];
     XCTAssertEqualObjects(bold, item, @"Expected %d but found %@", (int) normal, item);
 }
@@ -89,7 +89,7 @@
     NSArray *result = [PXFontEntry filterEntries:entries byStyle:@"normal"];
 
     XCTAssertNotNil(result, @"Expected an array back when filtering by style");
-    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %d", result.count);
+    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %lu", (unsigned long)result.count);
     PXFontEntry *item = result[0];
     XCTAssertEqualObjects(normal, item, @"Expected %@ but found %@", normal, item);
 }
@@ -104,7 +104,7 @@
     NSArray *result = [PXFontEntry filterEntries:entries byStyle:@"normal"];
 
     XCTAssertNotNil(result, @"Expected an array back when filtering by style");
-    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %d", result.count);
+    XCTAssertTrue(result.count == 1, @"Expected only a single font info but found %lu", (unsigned long)result.count);
     PXFontEntry *item = result[0];
     XCTAssertEqualObjects(italic, item, @"Expected %d but found %@", (int) normal, item);
 }
